requirements.txt
APScheduler==3.9.1
attrs==22.1.0
click==8.1.3
coverage==6.5.0
exceptiongroup==1.0.0
Flask==2.2.2
Flask-APScheduler==1.12.4
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.0
greenlet==1.1.3
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==1.4.41
tomli==2.0.1
tzdata==2022.5
tzlocal==4.2
Werkzeug==2.2.2
zipp==3.8.1
